extractd common RecipeCard component
