+++ meal-ui/src/components/Card/ImageCard.js
@@ -1,7 +1,15 @@
 import StatusBadge from 'components/StatusBadge';
 import { ClickableTile } from '@carbon/react';
+import { ArrowRight } from '@carbon/icons-react';
 
-const ImageCard = ({ id, className, label, status, ctaHandler, children }) => {
+const ImageCard = ({
+	id,
+	className,
+	label,
+	status,
+	ctaHandler,
+	...additionalProps
+}) => {
 	return (
 		<ClickableTile
 			onClick={() => ctaHandler(id)}
@@ -9,13 +17,15 @@ const ImageCard = ({ id, className, label, status, ctaHandler, children }) => {
 				className ? `card image-card ${className}` : `card image-card`
 			}
 		>
-			{children}
+			{additionalProps.children}
 			<StatusBadge
 				className='e-image-status-badge'
 				status={status}
 			/>
 			<div className='action-slot'>
 				<label className='action-label'>{label}</label>
+				Got to {additionalProps.entityName ?? ''}
+				<ArrowRight />
 			</div>
 		</ClickableTile>
 	);
