+++ meal-ui/src/pages/Recipes/RecipeGrid.jsx
@@ -1,0 +1,14 @@
+import { GridTile } from './GridTile';
+
+export function RecipeGrid({ recipes, addOrRemoveSelectedItem }) {
+	return (
+		<div className='recipes-grid scrollbar-vertical'>
+			{recipes.map(recipe => (
+				<GridTile
+					item={recipe}
+					gridHandler={addOrRemoveSelectedItem}
+				/>
+			))}
+		</div>
+	);
+}
